Rename HomePages to be coherent with other components' naming
