+++ speedy-recovery-client/src/__tests__/components/HomePageFactory.test.js
@@ -2,9 +2,9 @@ import React from "react";
 import Enzyme, { mount } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import HomePageFactory from "../../components/secured/home/HomePageFactory";
-import HomePageForPatient from "../../components/secured/home/HomePageForPatient";
-import HomePageForParent from "../../components/secured/home/HomePageForParent";
-import HomePageForPractitioner from "../../components/secured/home/HomePageForPractitioner";
+import PatientHomePage from "../../components/secured/home/PatientHomePage";
+import ParentHomePage from "../../components/secured/home/ParentHomePage";
+import PractitionerHomePage from "../../components/secured/home/PractitionerHomePage";
 
 Enzyme.configure({ adapter: new Adapter() });
 
@@ -27,7 +27,7 @@ test("HomePageFactory renders patient's home page for patient users", () => {
     />
   );
   expect(wrapper.children().length).toBe(1);
-  expect(wrapper.children().type().name).toEqual("HomePageForPatient");
+  expect(wrapper.children().type().name).toEqual("PatientHomePage");
 });
 
 test("HomePageFactory renders parent's home page for parent users", () => {
@@ -39,7 +39,7 @@ test("HomePageFactory renders parent's home page for parent users", () => {
     />
   );
   expect(wrapper.children().length).toBe(1);
-  expect(wrapper.children().type().name).toEqual("HomePageForParent");
+  expect(wrapper.children().type().name).toEqual("ParentHomePage");
 });
 
 test("HomePageFactory renders practitioner's home page for practitioner users", () => {
@@ -51,5 +51,5 @@ test("HomePageFactory renders practitioner's home page for practitioner users", 
     />
   );
   expect(wrapper.children().length).toBe(1);
-  expect(wrapper.children().type().name).toEqual("HomePageForPractitioner");
+  expect(wrapper.children().type().name).toEqual("PractitionerHomePage");
 });
